Changes to get unit tests to run
The //test/common/buffer/... tests pass on Windows

Tests under //test/common/event pass on Windows

- libevent does not support EV_ET on windows, so we make it an error to
use FileTriggerType::Edge on Windows
Add "all_windows" target for each group of tests we get to pass

Consolidate some "#undef" statements

Tests under //test/common/network/... pass on Windows
Tests under //test/common/tcp/... pass on Windows
Tests under //test/common/tcp_proxy/... pass on Windows
Add some more tests that already pass on Windows

See https://docs.google.com/document/d/1PlxrRNri0N7_iofOmj4IQ7XljnWiIZ5GOC_tvVc-C2k/edit
for details

Tests under //test/common/ssl/... pass on Windows

- runfiles for tests are not symlinked, so we need to look up their
location in a MANIFEST file

- On Windows, when two listeners use the same cert and session ticket key, but
different client CA, the session cannot be resumed between the two. This is the correct
behavior. However, the exact error the server gets is slightly different
(SSL_ERROR_SYSCALL on Windows vs. SSL_ERROR_SSL with TLSV1_ALERT_UNKNOWN_CA on Linux)
The tests are updated to reflect this

Tests under //test/common/api/... pass on Windows
Tests under //test/common/compressor/... pass on Windows
Tests under //test/common/json/... pass on Windows
Tests under //test/common/request_info/... pass on Windows
Tests under //test/common/secret/... pass on Windows
//test/common/filesystem:filesystem_impl_test passes on Windows

- we needed to add separate methods for writeFile / writeSocket and
closeFile / closeSocket to handle the fact that files and sockets have
different types on windows (int vs. SOCKET)

//test/integration:tcp_proxy_integration_test passes on Windows
//test/integration:http2_integration_test passes on Windows
//test/integration:integration_test passes on Windows
//test/integration:http2_upstream_integration_test passes on Windows
//test/integration:ssl_integration_test passes on Windows

Pass EVLOOP_NONBLOCK | EVLOOP_ONCE to event_base_loop.
This results in the expected behavior for NonBlock

Ensure the capture transport socket writes binary files in binary mode

Route all syscalls through Api::OsSysCallsImpl

Split os_sys_calls implementation into linux and win32 packages

Enable tcp_proxy filter on Windows

Add implementation for Win32 filewatcher

This uses the ReadDirectoryChangesW API. Since this API does not work at
all with libevent, use a socket on localhost and a conccurrent queue to
synchronize the execution of the callbacks

Add ability to build release binaries on Windows

Change nghttp2 static lib name to nghttp2_static

This is because their build system also produces an nghttp2.lib file
that dynamically links to nghttp2.dll; we do not want to use that one

Run tests under test_common on windows

Run redis_proxy extension tests on Windows

Split thread impl into separate win32 and pthread

+++ source/common/upstream/upstream_impl.h
@@ -206,7 +206,7 @@ protected:
                    const Network::ConnectionSocket::OptionsSharedPtr& options);
 
 private:
-  std::atomic<uint64_t> health_flags_{};
+  std::atomic<uint32_t> health_flags_{};
   ActiveHealthFailureType active_health_failure_type_{};
   std::atomic<uint32_t> weight_;
   std::atomic<bool> used_;
